requirements.txt
pandas-gbq
apache-airflow
pydantic-settings